# pkg/server/docs/swagger.yaml
basePath: /
definitions:
  entities.EndpointsInfo:
    properties:
      all:
        type: string
      latin:
        type: string
      urutan:
        type: string
    type: object
  models.APIErrorResponse:
    properties:
      message:
        type: string
      stack:
        type: string
      statusCode:
        type: integer
    type: object
  models.APIInfoResponse:
    properties:
      author:
        type: string
      endpoints:
        $ref: '#/definitions/entities.EndpointsInfo'
      repository:
        type: string
    type: object
  models.APISuccessResponse:
    properties:
      data:
        $ref: '#/definitions/responses.AsmaulHusna'
      statusCode:
        type: integer
      total:
        type: integer
    type: object
  responses.AsmaulHusna:
    properties:
      arab:
        type: string
      arti:
        type: string
      latin:
        type: string
      urutan:
        type: integer
    type: object
host: https://asmaul-husna-api.vercel.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Asmaul Husna API
  license:
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Asmaul Husna API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Get Info about the API.
      produces:
      - application/json
      responses:
        "200":
          description: Get Info success!
          schema:
            $ref: '#/definitions/models.APIInfoResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIErrorResponse'
      summary: Get Info about the API
      tags:
      - AsmaulHusna
  /api/{urutan}:
    get:
      consumes:
      - application/json
      description: Get Asma'ul Husna based on urutan
      parameters:
      - description: Urutan
        in: path
        name: urutan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get Asma'ul Husna based on urutan success!
          schema:
            $ref: '#/definitions/models.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APISuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIErrorResponse'
      summary: Get Asma'ul Husna based on urutan
      tags:
      - AsmaulHusna
  /api/all:
    get:
      consumes:
      - application/json
      description: Get all Asma'ul Husna with optional query page and limit
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get Asma'ul Husna success!
          schema:
            $ref: '#/definitions/models.APISuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIErrorResponse'
      summary: Get all Asma'ul Husna with optional query page and limit
      tags:
      - AsmaulHusna
  /api/latin/{latin}:
    get:
      consumes:
      - application/json
      description: Get Asma'ul Husna based on latin.
      parameters:
      - description: Latin
        in: path
        name: latin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get Asma'ul Husna based on latin success!
          schema:
            $ref: '#/definitions/models.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIErrorResponse'
      summary: Get Asma'ul Husna based on latin
      tags:
      - AsmaulHusna
  /metrics:
    get:
      consumes:
      - application/json
      description: Get Metrics
      produces:
      - application/json
      responses:
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Metrics
      tags:
      - Metrics
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"